feat: Implement Browser History functionality
- Added `browserHistory.h` with Browser History object definition and its function prototypes.
- Added `browserHistory.c` with function implementations to manipulate Browser History objects.
- Added functions to create and destroy Browser History objects, and push and delete Browser History Item objects from them.

+++ browserHistoryItem.h
@@ -4,9 +4,8 @@
 #include "date.h"
 #include "webPage.h"
 
-typedef struct BrowserHistoryItem
-{
-    short unsigned int browserHistoryItemId;
+typedef struct BrowserHistoryItem {
+    unsigned int browserHistoryItemId;
     WebPage* pWebPage;
     Date dateVisited;
     struct BrowserHistoryItem* pNext;
